% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCluster-package.R
\docType{package}
\name{autoCluster-package}
\alias{autoCluster}
\alias{autoCluster-package}
\title{autoCluster: Automatic Clustering}
\description{
Performs unsupervised clustering and automatically
    determine the best number of cluster.
}
\author{
\strong{Maintainer}: Etienne Camenen \email{etienne.camenen@gmail.com}

Authors:
\itemize{
  \item Maxime Chazalviel
  \item Fabien Jourdan
}

}
\keyword{internal}
